# [Fix #10919] Fix a huge performance regression between 1.32.0 and 1.33.0