--- a/changelog/fix_fix_a_huge_performance_regression.md
+++ b/changelog/fix_fix_a_huge_performance_regression.md
@@ -1,0 +1,1 @@
+* [#10919](https://github.com/rubocop/rubocop/issues/10919): Fix a huge performance regression between 1.32.0 and 1.33.0. ([@ydah][])