@dvandersluis 2022-08-24
Good catch! Would any other cops changed in #10829 be affected?