@ydah 2022-08-24
@dvandersluis Yes. I think so. 
the parameters obsoleted in https://github.com/rubocop/rubocop/pull/10555 had the same problem.